+++ crates/zng-wgt-material-icons/build.rs
@@ -1,34 +1,46 @@
 use std::{env, error::Error, fmt, fs, path::PathBuf};
 
 static FONTS: &[(&str, &str, &str, &[u8], &str)] = &[
-    (
-        "outlined",
-        "otf",
-        "opentype",
-        include_bytes!("fonts/MaterialIconsOutlined-Regular.otf"),
-        include_str!("fonts/MaterialIconsOutlined-Regular.codepoints"),
-    ),
-    (
-        "filled",
-        "ttf",
-        "truetype",
-        include_bytes!("fonts/MaterialIcons-Regular.ttf"),
-        include_str!("fonts/MaterialIcons-Regular.codepoints"),
-    ),
-    (
-        "rounded",
-        "otf",
-        "opentype",
-        include_bytes!("fonts/MaterialIconsRound-Regular.otf"),
-        include_str!("fonts/MaterialIconsRound-Regular.codepoints"),
-    ),
-    (
-        "sharp",
-        "otf",
-        "opentype",
-        include_bytes!("fonts/MaterialIconsSharp-Regular.otf"),
-        include_str!("fonts/MaterialIconsSharp-Regular.codepoints"),
-    ),
+    #[cfg(feature = "outlined")]
+    {
+        (
+            "outlined",
+            "otf",
+            "opentype",
+            include_bytes!("fonts/MaterialIconsOutlined-Regular.otf"),
+            include_str!("fonts/MaterialIconsOutlined-Regular.codepoints"),
+        )
+    },
+    #[cfg(feature = "filled")]
+    {
+        (
+            "filled",
+            "ttf",
+            "truetype",
+            include_bytes!("fonts/MaterialIcons-Regular.ttf"),
+            include_str!("fonts/MaterialIcons-Regular.codepoints"),
+        )
+    },
+    #[cfg(feature = "rounded")]
+    {
+        (
+            "rounded",
+            "otf",
+            "opentype",
+            include_bytes!("fonts/MaterialIconsRound-Regular.otf"),
+            include_str!("fonts/MaterialIconsRound-Regular.codepoints"),
+        )
+    },
+    #[cfg(feature = "sharp")]
+    {
+        (
+            "sharp",
+            "otf",
+            "opentype",
+            include_bytes!("fonts/MaterialIconsSharp-Regular.otf"),
+            include_str!("fonts/MaterialIconsSharp-Regular.codepoints"),
+        )
+    },
 ];
 
 fn main() {
@@ -40,59 +52,47 @@ fn main() {
 }
 
 fn write(codepoints: &str, mod_name: &str) {
-    let code = generate(codepoints, mod_name).unwrap();
+    let [docs, map] = generate(codepoints, mod_name).unwrap();
 
-    let generated = PathBuf::from(env::var("OUT_DIR").unwrap()).join(format!("generated.{mod_name}.rs"));
-    fs::write(generated, code.as_bytes()).unwrap();
+    let generated = PathBuf::from(env::var("OUT_DIR").unwrap()).join(format!("generated.{mod_name}.map.rs"));
+    fs::write(generated, map.as_bytes()).unwrap();
+
+    let generated = PathBuf::from(env::var("OUT_DIR").unwrap()).join(format!("generated.{mod_name}.docs.txt"));
+    fs::write(generated, docs.as_bytes()).unwrap();
 }
-fn generate(codepoints: &str, mod_name: &str) -> Result<String, Box<dyn Error>> {
+fn generate(codepoints: &str, mod_name: &str) -> Result<[String; 2], Box<dyn Error>> {
     use fmt::Write;
 
-    let mut s = String::new();
+    let mut docs = String::new();
+    let mut map = phf_codegen::Map::new();
 
-    let mut all = vec![];
+    let mut buffer = String::with_capacity(3);
     for line in codepoints.lines() {
         if let Some((name, code)) = line.split_once(' ') {
             if name.is_empty() || code.is_empty() {
                 return Err("invalid codepoints file".into());
             }
 
-            let name = if name.chars().next().unwrap().is_ascii_digit() {
-                format!("N{name}").to_uppercase()
-            } else {
-                name.to_uppercase()
-            };
-
             let code = u32::from_str_radix(code, 16)?;
             let code = char::from_u32(code).ok_or("invalid codepoint")?;
+            buffer.push('\'');
+            buffer.push(code);
+            buffer.push('\'');
 
-            writeln!(&mut s)?;
-            writeln!(&mut s, r#"/// <span class="material-icons {mod_name}">{code}</span>"#)?;
-            writeln!(&mut s, r#"/// "#)?;
-            writeln!(&mut s, r#"/// <span class="material-icons large {mod_name}">{code}</span>"#)?;
-            writeln!(
-                &mut s,
-                r#"pub const {name}: MaterialIcon = MaterialIcon {{ font: meta::FONT_NAME, name: "{name}", code: '{code}', }};"#
-            )?;
+            writeln!(&mut docs, r#"| {name} | <span class="material-icons {mod_name}">{code}</span> |"#)?;
 
-            all.push(name);
+            map.entry(name, &buffer);
+            buffer.clear();
         } else {
             return Err("invalid codepoints file".into());
         }
     }
+    let map = format!(
+        "/// Map of name to icon codepoint.\n\n\npub static MAP: phf::Map<&'static str, char> = {};",
+        map.build()
+    );
 
-    writeln!(&mut s)?;
-    writeln!(&mut s, "/// All icons.")?;
-    writeln!(&mut s, "pub fn all() -> Vec<MaterialIcon> {{")?;
-    write!(&mut s, "    vec![")?;
-    for name in all {
-        write!(&mut s, "{name}")?;
-        write!(&mut s, ", ")?;
-    }
-    writeln!(&mut s, "]")?;
-    writeln!(&mut s, "}}")?;
-
-    Ok(s)
+    Ok([docs, map])
 }
 
 fn write_html_in_header() {
@@ -105,9 +105,6 @@ fn write_html_in_header() {
 
     use std::fmt::Write;
     let mut css = String::new();
-    writeln!(&mut css, ".material-icons.large {{").unwrap();
-    writeln!(&mut css, "   font-size: 32px;").unwrap();
-    writeln!(&mut css, "}}").unwrap();
 
     for (mod_name, ext, format, font, _) in FONTS {
         let mut file = doc_dir.join(mod_name);
@@ -124,6 +121,7 @@ fn write_html_in_header() {
         writeln!(&mut css, "}}").unwrap();
         writeln!(&mut css, ".material-icons.{mod_name} {{").unwrap();
         writeln!(&mut css, "   font-family: \"zng-material-icons-extensions-{mod_name}\";").unwrap();
+        writeln!(&mut css, "   font-size: 32px;").unwrap();
         writeln!(&mut css, "}}").unwrap();
     }
     fs::write(&file, css).unwrap();