+++ crates/zng/src/lib.rs
@@ -628,7 +628,7 @@ mod __prelude {
         WINDOWS,
     };
 
-    pub use zng_wgt::Wgt;
+    pub use zng_wgt::{CommandIconExt as _, Wgt, ICONS};
 
     pub use crate::text;
     pub use zng_wgt_text::Text;
@@ -649,8 +649,6 @@ mod __prelude {
 
     pub use crate::layer;
     pub use zng_wgt_layer::{AnchorMode, LayerIndex, LAYERS};
-
-    pub use zng_wgt_text::icon::CommandIconExt as _;
 
     pub use crate::popup;
     pub use zng_wgt_layer::popup::POPUP;
@@ -850,9 +848,9 @@ mod defaults {
         /// * [`UndoManager`]
         /// * [`SingleInstanceManager`] if the `"single_instance"` feature is enabled.
         /// * [`HotReloadManager`] if the `"hot_reload"` feature is enabled.
-        /// * [`MaterialFonts`] if any `"material_icons*"` feature is enabled.
+        /// * [`MaterialIconsManager`] if any `"material_icons*"` feature is enabled.
         ///
-        /// [`MaterialFonts`]: zng_wgt_material_icons::MaterialFonts
+        /// [`MaterialIconsManager`]: zng_wgt_material_icons::MaterialIconsManager
         /// [`SingleInstanceManager`]: zng_ext_single_instance::SingleInstanceManager
         /// [`HotReloadManager`]: zng_ext_hot_reload::HotReloadManager
         /// [`ConfigManager`]: zng_ext_config::ConfigManager
@@ -891,7 +889,7 @@ mod defaults {
                 feature = "material_icons_rounded",
                 feature = "material_icons_sharp",
             ))]
-            let r = r.extend(zng_wgt_material_icons::MaterialFonts);
+            let r = r.extend(zng_wgt_material_icons::MaterialIconsManager);
 
             r.extend(DefaultsInit {})
         }
@@ -946,55 +944,6 @@ mod defaults {
                     }))
                     .perm();
                 tracing::debug!("defaults init, single_instance set");
-            }
-
-            #[cfg(feature = "material_icons_outlined")]
-            {
-                use zng_app::widget::node::{NilUiNode, UiNode};
-                use zng_ext_clipboard::*;
-                use zng_ext_undo::*;
-                use zng_ext_window::cmd::*;
-                use zng_wgt::{wgt_fn, IconRequestArgs, ICONS};
-                use zng_wgt_input::cmd::*;
-                use zng_wgt_material_icons::outlined as icons;
-                use zng_wgt_scroll::cmd::*;
-                use zng_wgt_text::icon::CommandIconExt as _;
-                use zng_wgt_text::icon::Icon;
-
-                CUT_CMD.init_icon(wgt_fn!(|_| Icon!(icons::CUT)));
-                COPY_CMD.init_icon(wgt_fn!(|_| Icon!(icons::COPY)));
-                PASTE_CMD.init_icon(wgt_fn!(|_| Icon!(icons::PASTE)));
-
-                UNDO_CMD.init_icon(wgt_fn!(|_| Icon!(icons::UNDO)));
-                REDO_CMD.init_icon(wgt_fn!(|_| Icon!(icons::REDO)));
-
-                CLOSE_CMD.init_icon(wgt_fn!(|_| Icon!(icons::CLOSE)));
-                MINIMIZE_CMD.init_icon(wgt_fn!(|_| Icon!(icons::MINIMIZE)));
-                MAXIMIZE_CMD.init_icon(wgt_fn!(|_| Icon!(icons::MAXIMIZE)));
-                FULLSCREEN_CMD.init_icon(wgt_fn!(|_| Icon!(icons::FULLSCREEN)));
-
-                CONTEXT_MENU_CMD.init_icon(wgt_fn!(|_| Icon!(icons::MENU_OPEN)));
-
-                #[cfg(feature = "inspector")]
-                zng_wgt_inspector::INSPECT_CMD.init_icon(wgt_fn!(|_| Icon!(icons::SCREEN_SEARCH_DESKTOP)));
-
-                SCROLL_TO_TOP_CMD.init_icon(wgt_fn!(|_| Icon!(icons::VERTICAL_ALIGN_TOP)));
-                SCROLL_TO_BOTTOM_CMD.init_icon(wgt_fn!(|_| Icon!(icons::VERTICAL_ALIGN_BOTTOM)));
-
-                ZOOM_IN_CMD.init_icon(wgt_fn!(|_| Icon!(icons::ZOOM_IN)));
-                ZOOM_OUT_CMD.init_icon(wgt_fn!(|_| Icon!(icons::ZOOM_OUT)));
-
-                OPEN_CMD.init_icon(wgt_fn!(|_| Icon!(icons::FILE_OPEN)));
-                SETTINGS_CMD.init_icon(wgt_fn!(|_| Icon!(icons::SETTINGS)));
-
-                ICONS.register(wgt_fn!(|args: IconRequestArgs| {
-                    match args.name() {
-                        "settings-reset" => Icon!(icons::SETTINGS_BACKUP_RESTORE).boxed(),
-                        _ => NilUiNode.boxed(),
-                    }
-                }));
-
-                tracing::debug!("defaults init, command_icons set");
             }
         }
     }