+++ examples/icon/src/main.rs
@@ -1,4 +1,4 @@
-//! Search and copy Material Icons constants.
+//! Search and copy Material Icons keys.
 
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
@@ -9,11 +9,12 @@ use zng::{
         color_scheme_map,
         filter::{backdrop_blur, drop_shadow, opacity},
     },
+    container,
     data_view::{DataView, DataViewArgs},
     focus::{directional_nav, focus_scope, focus_shortcut, tab_nav, DirectionalNav, TabNav},
     font::FontName,
     gesture::{on_click, ClickArgs},
-    icon::{self, Icon, MaterialIcon},
+    icon::{self, GlyphIcon, Icon},
     label::Label,
     layout::{align, margin, padding},
     prelude::*,
@@ -29,7 +30,7 @@ fn main() {
         Window! {
             title = "Icon Example";
             icon = WindowIcon::render(|| Icon! {
-                ico = icon::material_filled::LIGHTBULB;
+                ico = icon::material::filled::req("lightbulb");
                 ico_color = colors::YELLOW;
                 drop_shadow = (0, 0), 3, colors::WHITE;
             });
@@ -52,7 +53,7 @@ fn icons() -> impl UiNode {
             value::<&'static str> = key;
         }
     }
-    fn show_font(icons: Vec<MaterialIcon>, font_mod: &'static str) -> impl UiNode {
+    fn show_font(icons: Vec<(&'static str, GlyphIcon)>, font_mod: &'static str) -> impl UiNode {
         let _scope = tracing::error_span!("show_font").entered();
         let icons_len = icons.len();
         Wrap! {
@@ -75,7 +76,7 @@ fn icons() -> impl UiNode {
                     children = {
                         let mut r = vec![];
                         c.par_iter()
-                                .map(|i| icon_btn(i.clone(), font_mod).boxed())
+                                .map(|(name, ico)| icon_btn(name, ico.clone(), font_mod).boxed())
                                 .collect_into_vec(&mut r);
                         r
                     };
@@ -113,7 +114,7 @@ fn icons() -> impl UiNode {
                 focus_shortcut = [shortcut!['S'], shortcut![CTRL+'F'], shortcut![Find]];
                 widget::foreground = Icon! {
                     align = Align::LEFT;
-                    ico = icon::material_outlined::SEARCH;
+                    ico = icon::material::outlined::req("search");
                     ico_size = 18;
                     margin = (0, 0, 0, 6);
                 };
@@ -143,23 +144,23 @@ fn icons() -> impl UiNode {
                 merge_var!(selected_font, search, |f, s| (*f, s.clone())),
                 hn!(|a: &DataViewArgs<(&'static str, Txt)>| {
                     if let Some((f, s)) = a.get_new() {
-                        let mut icons = match f {
-                            "filled" => icon::material_filled::all(),
-                            "outlined" => icon::material_outlined::all(),
-                            "rounded" => icon::material_rounded::all(),
-                            "sharp" => icon::material_sharp::all(),
+                        let mut icons: Vec<_> = match f {
+                            "filled" => icon::material::filled::all().collect(),
+                            "outlined" => icon::material::outlined::all().collect(),
+                            "rounded" => icon::material::rounded::all().collect(),
+                            "sharp" => icon::material::sharp::all().collect(),
                             _ => unreachable!(),
                         };
+                        icons.sort_by_key(|(k, _)| *k);
                         if let Some(len) = s.strip_prefix("-len") {
                             let len: usize = len.trim().parse().unwrap_or(0);
                             icons.retain(|f| {
-                                f.name.len() >= len
+                                f.0.len() >= len
                             });
                         } else if !s.is_empty() {
-                            let s = s.to_uppercase();
+                            let s = s.to_lowercase();
                             icons.retain(|f| {
-                                f.name.contains(&s) ||
-                                f.display_name().to_uppercase().contains(&s)
+                                f.0.contains(&s)
                             });
                         }
                         if icons.is_empty() {
@@ -177,7 +178,7 @@ fn icons() -> impl UiNode {
     }
 }
 
-fn icon_btn(ico: icon::MaterialIcon, font_mod: &'static str) -> impl UiNode {
+fn icon_btn(name: &'static str, ico: icon::GlyphIcon, font_mod: &'static str) -> impl UiNode {
     Button! {
         padding = 2;
         layout::size = (80, 80);
@@ -190,7 +191,7 @@ fn icon_btn(ico: icon::MaterialIcon, font_mod: &'static str) -> impl UiNode {
                     ico = ico.clone();
                 },
                 Text! {
-                    txt = formatx!("{ico}");
+                    txt = name;
                     txt_align = Align::CENTER;
                     font_size = 10;
                     layout::height = 2.em();
@@ -199,12 +200,12 @@ fn icon_btn(ico: icon::MaterialIcon, font_mod: &'static str) -> impl UiNode {
             ]
         };
         on_click = hn!(|_| {
-            LAYERS.insert(LayerIndex::TOP_MOST, expanded_icon(ico.clone(), font_mod));
+            LAYERS.insert(LayerIndex::TOP_MOST, expanded_icon(name, ico.clone(), font_mod));
         })
     }
 }
 
-fn expanded_icon(ico: icon::MaterialIcon, font_mod: &'static str) -> impl UiNode {
+fn expanded_icon(name: &'static str, ico: icon::GlyphIcon, font_mod: &'static str) -> impl UiNode {
     let opacity = var(0.fct());
     opacity.ease(1.fct(), 200.ms(), easing::linear).perm();
     Container! {
@@ -235,33 +236,10 @@ fn expanded_icon(ico: icon::MaterialIcon, font_mod: &'static str) -> impl UiNode
                     padding = 10;
                     children_align = Align::TOP_LEFT;
                     children = ui_vec![
-                        title(formatx!("{ico}")),
-                        {
-                            let full_path = formatx!("icon::{font_mod}::{}", ico.name);
-                            let copied = var(false);
-                            Label! {
-                                txt = ico.name;
-                                font_family = FontName::monospace();
-                                font_size = 18;
-                                mouse::cursor = mouse::CursorIcon::Pointer;
-                                widget::enabled = copied.map(|&c| !c);
-                                tooltip = Tip!(Text!("copy '{full_path}'"));
-                                tip::disabled_tooltip = Tip!(Text!("copied!"));
-                                on_click = async_hn!(copied, full_path, |_| {
-                                    if clipboard::CLIPBOARD.set_text(full_path).wait_rsp().await.is_ok() {
-                                        ACCESS.show_tooltip(WINDOW.id(), WIDGET.id());
-                                        copied.set(true);
-                                        task::deadline(2.secs()).await;
-                                        copied.set(false);
-                                    }
-                                });
-                                when *#gesture::is_hovered {
-                                    background_color = text::FONT_COLOR_VAR.map(|c| c.with_alpha(20.pct()));
-                                }
-                            }
-                        },
-                        sub_title("Using `Icon!`:"),
+                        title(name.into()),
                         Stack! {
+                            align = Align::CENTER;
+                            margin = 10;
                             direction = StackDirection::left_to_right();
                             spacing = 5;
                             children_align = Align::TOP_LEFT;
@@ -286,40 +264,14 @@ fn expanded_icon(ico: icon::MaterialIcon, font_mod: &'static str) -> impl UiNode
                                 }.boxed()
                             })).collect::<Vec<_>>()
                         },
-
-                        sub_title("Using `Text!`:"),
-                        Stack! {
-                            direction = StackDirection::left_to_right();
-                            spacing = 5;
-                            children_align = Align::TOP_LEFT;
-                            children = [64, 48, 32, 24, 16].into_iter().map(clmv!(ico, |size| {
-                                Stack! {
-                                    direction = StackDirection::top_to_bottom();
-                                    spacing = 3;
-                                    children = ui_vec![
-                                        size_label(formatx!("{size}")),
-                                        Text! {
-                                            txt = ico.code;
-                                            font_family = ico.font.clone();
-                                            font_size = size;
-
-                                            background_color = color_scheme_map(
-                                                colors::BLACK.with_alpha(85.pct()),
-                                                colors::WHITE.with_alpha(85.pct())
-                                            );
-                                            corner_radius = 4;
-                                            padding = 2;
-                                        }
-                                    ]
-                                }.boxed()
-                            })).collect::<Vec<_>>()
-                        }
+                        code_copy("ICONS.req".into(), formatx!("ICONS.req(\"material/{font_mod}/{name}\")")),
+                        code_copy(formatx!("{font_mod}::req"), formatx!("icon::{font_mod}::req(\"{name}\")")),
                     ]
                 },
                 Button! {
                     id = "close-btn";
                     icon::ico_size = 14;
-                    child = Icon!(icon::material_filled::CLOSE);
+                    child = Icon!(icon::material::filled::req("close"));
                     align = Align::TOP_RIGHT;
                     padding = 2;
                     margin = 4;
@@ -344,16 +296,35 @@ fn title(title: Txt) -> impl UiNode {
         txt_align = Align::CENTER;
     }
 }
-fn sub_title(title: impl Into<Txt>) -> impl UiNode {
-    Text! {
-        txt = title.into();
-        font_size = 16;
-    }
-}
+
 fn size_label(size: Txt) -> impl UiNode {
     Text! {
         txt = size;
         font_size = 10;
         txt_align = Align::CENTER;
+    }
+}
+
+fn code_copy(label: Txt, code: Txt) -> impl UiNode {
+    let copied = var(false);
+    Label! {
+        txt = label;
+        font_family = FontName::monospace();
+        mouse::cursor = mouse::CursorIcon::Pointer;
+        widget::enabled = copied.map(|&c| !c);
+        tooltip = Tip!(Text!("copy {code}"));
+        tip::disabled_tooltip = Tip!(Text!("copied!"));
+        container::child_start = ICONS.get("copy"), 4;
+        on_click = async_hn!(copied, code, |_| {
+            if clipboard::CLIPBOARD.set_text(code).wait_rsp().await.is_ok() {
+                ACCESS.show_tooltip(WINDOW.id(), WIDGET.id());
+                copied.set(true);
+                task::deadline(2.secs()).await;
+                copied.set(false);
+            }
+        });
+        when *#gesture::is_hovered {
+            background_color = text::FONT_COLOR_VAR.map(|c| c.with_alpha(20.pct()));
+        }
     }
 }